BREAKING CHANGE: udkc no longer interprets process.argv as concatenated config path
Before, udkc gives `process.argv.slice(2)` as devContainer args.
The devContainer function binds createContainer and this last concatenates
arguments (except the first which is the runtime path) to determine the
config path.
In that way, the process.argv isn't usable in the container runtime.
We must call devContainer programmatically to bypass udkc.

Now, udkc gives `process.env.UDKC` only as devContainer args.
So it's still possible to define a config path with the CLI's
environment variable `UDKC`.